+++ Source/AIM/StatsGenerator/SGPowerLawTableModel.h
@@ -86,11 +86,11 @@ Q_OBJECT    ;
 
     QAbstractItemDelegate* getItemDelegate();
 
-    void setBinNumbers(QVector<int> binNumbers);
+    void setBinNumbers(QVector<double> binNumbers);
 
-    QVector<qint32>& getBinNumbers()
+    QVector<double>& getBinNumbers()
     { return m_BinNumbers;}
-    qint32 getBinNumber(qint32 row)
+    double getBinNumber(qint32 row)
     { return m_BinNumbers[row];}
 
     QVector<QString>& getColors()
@@ -119,7 +119,7 @@ Q_OBJECT    ;
     int m_ColumnCount;
     int m_RowCount;
 
-    QVector<qint32> m_BinNumbers;
+    QVector<double> m_BinNumbers;
     QVector<double> m_Alpha;
     QVector<double> m_K;
     QVector<double> m_Beta;
